internship_bot_v7_production/internship_bot/requirements.txt
anthropic>=0.28.0
python-dotenv>=1.0.0
pydantic>=2.7.0
pydantic-settings>=2.0.0
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
celery[redis]>=5.4.0
redis>=5.0.0
flower>=2.0.1
spacy>=3.7.0
rapidfuzz>=3.9.0
nltk>=3.8.0
scikit-learn>=1.4.0
pdfminer.six>=20221105
pypdf>=4.2.0
python-docx>=1.1.0
weasyprint>=62.0
playwright>=1.44.0
sqlalchemy>=2.0.30
alembic>=1.13.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.0
httpx>=0.27.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.2.0
imapclient>=3.0.1
cryptography>=42.0.0
tenacity>=8.3.0
structlog>=24.1.0
rich>=13.7.0
typer>=0.12.0
jinja2>=3.1.0
pytest>=8.2.0
pytest-asyncio>=0.23.0
pytest-mock>=3.14.0
httpx[test]>=0.27.0
ruff>=0.4.0
mypy>=1.10.0
prometheus-client>=0.20.0
python-jose[cryptography]>=3.3.0   # JWT tokens
passlib[bcrypt]>=1.7.4             # password hashing
pip-audit>=2.7.0                   # CVE scanning (dev only)
aiosqlite>=0.20.0                  # async SQLite for dev/testing
python-multipart
